requirements.txt
yfinance
pandas
streamlit
polygon-api-client
